Simplify core crate
